+++ README.md
@@ -1,2 +1,4 @@
 # caixaEletronico_simulador
 Repositório para fins acadêmicos, que simula o comportmento de um caixa eletrônico e foi desenvolvido usando a liguagem Python.
+# Pull Request
+Nesse caso estamos realizando a atividade 2.
