+++ atm.py
@@ -3,7 +3,7 @@
 #Exemplo 1: Para sacar a quantia de 256 reais, o programa fornece duas notas de 100, uma nota de 50, uma nota de 5 e uma nota de 1;​
 #Exemplo 2: Para sacar a quantia de 399 reais, o programa fornece três notas de 100, uma nota de 50, quatro notas de 10,
 #uma nota de 5 e quatro notas de 1.
-from codigo_python import *
+from printar_notas import *
 
 dinheiro = int(input("Informe o valor a ser sacado: R$"))
 print("Para sacar a quantia de R$ {}, será necessário:".format(dinheiro))
@@ -13,5 +13,5 @@ if dinheiro > 10 and dinheiro <= 600:
     notas_10 = int(((dinheiro % 100) % 50)/ 10)
     notas_5 = int((((dinheiro % 100) % 50) % 10)/5)
     moeda_1 = int(dinheiro - (notas_100*100 + notas_50*50 + notas_10*10 + notas_5*5))
-
+#Chamando função no arquivo printar_notas
 printar_atm(notas_100, notas_50, notas_10, notas_5, moeda_1)
